Refactor: 작가 등록 신청 기능 리팩토링
