+++ src/main/java/com/onetuks/goguma_bookstore/author/repository/AuthorJpaRepository.java
@@ -1,0 +1,6 @@
+package com.onetuks.goguma_bookstore.author.repository;
+
+import com.onetuks.goguma_bookstore.author.model.Author;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorJpaRepository extends JpaRepository<Author, Long> {}
