+++ src/main/java/com/onetuks/goguma_bookstore/author/service/dto/param/AuthorCreateParam.java
@@ -1,0 +1,3 @@
+package com.onetuks.goguma_bookstore.author.service.dto.param;
+
+public record AuthorCreateParam(String nickname, String introduction) {}
