--- src/main/java/com/onetuks/goguma_bookstore/author_debut/controller/dto/response/AuthorDebutCreateResponse.java
@@ -1,10 +1,0 @@
-package com.onetuks.goguma_bookstore.author_debut.controller.dto.response;
-
-import com.onetuks.goguma_bookstore.author_debut.service.dto.result.AuthorDebutCreateResult;
-
-public record AuthorDebutCreateResponse(Long memberId) {
-
-  public static AuthorDebutCreateResponse from(AuthorDebutCreateResult result) {
-    return new AuthorDebutCreateResponse(result.memberId());
-  }
-}
