--- src/main/java/com/onetuks/goguma_bookstore/author_debut/service/dto/param/AuthorDebutCreateParam.java
@@ -1,3 +1,0 @@
-package com.onetuks.goguma_bookstore.author_debut.service.dto.param;
-
-public record AuthorDebutCreateParam(String nickname, String introduction) {}
