+++ src/main/java/com/onetuks/goguma_bookstore/book/model/Book.java
@@ -2,7 +2,7 @@ package com.onetuks.goguma_bookstore.book.model;
 
 import static jakarta.persistence.CascadeType.REMOVE;
 
-import com.onetuks.goguma_bookstore.author_debut.model.Author;
+import com.onetuks.goguma_bookstore.author.model.Author;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
